+++ src/Splunk.OpenTelemetry.AutoInstrumentation/Traces.cs
@@ -42,7 +42,7 @@ internal class Traces
     public TracerProviderBuilder ConfigureTracerProvider(TracerProviderBuilder builder)
     {
         ServiceNameWarning.Instance.SendOnMissingServiceName(_log);
-        return builder.ConfigureResource(ResourceConfigurator.Configure);
+        return builder.ConfigureResource(b => ResourceConfigurator.Configure(b, _settings));
     }
 
     public void ConfigureTracesOptions(OtlpExporterOptions options)